refactor: simd swar
Moves the block-wise validators to a "swar" SIMD backend

The core logic of validate => extract => chain is now more evident
