+++ src/simd/runtime.rs
@@ -30,13 +30,17 @@ fn get_runtime_feature() -> u8 {
     feature
 }
 
+pub fn match_header_name_vectored(bytes: &mut Bytes) {
+    super::swar::match_header_name_vectored(bytes);
+}
+
 pub fn match_uri_vectored(bytes: &mut Bytes) {
     // SAFETY: calls are guarded by a feature check
     unsafe {
         match get_runtime_feature() {
             AVX2 => avx2::match_uri_vectored(bytes),
             SSE42 => sse42::match_uri_vectored(bytes),
-            _ => {},
+            _ /* NOP */ => super::swar::match_uri_vectored(bytes),
         }
     }
 }
@@ -47,7 +51,7 @@ pub fn match_header_value_vectored(bytes: &mut Bytes) {
         match get_runtime_feature() {
             AVX2 => avx2::match_header_value_vectored(bytes),
             SSE42 => sse42::match_header_value_vectored(bytes),
-            _ => {},
+            _ /* NOP */ => super::swar::match_header_value_vectored(bytes),
         }
     }
 }
